+++ src/main/java/com/fasterxml/jackson/datatype/threetenbp/deser/LocalTimeDeserializer.java
@@ -20,7 +20,9 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.JsonDeserializer;
 import org.threeten.bp.LocalTime;
+import org.threeten.bp.format.DateTimeFormatter;
 
 import java.io.IOException;
 
@@ -30,15 +32,24 @@ import java.io.IOException;
  * @author Nick Williams
  * @since 2.4.1
  */
-public class LocalTimeDeserializer extends ThreeTenDeserializerBase<LocalTime>
+public class LocalTimeDeserializer extends ThreeTenDateTimeDeserializerBase<LocalTime>
 {
     private static final long serialVersionUID = 1L;
 
     public static final LocalTimeDeserializer INSTANCE = new LocalTimeDeserializer();
 
-    private LocalTimeDeserializer()
-    {
-        super(LocalTime.class);
+    private LocalTimeDeserializer() {
+        this(DateTimeFormatter.ISO_LOCAL_TIME);
+    }
+
+    protected LocalTimeDeserializer(DateTimeFormatter dtf) {
+        super(LocalTime.class, dtf);
+
+    }
+
+    @Override
+    protected JsonDeserializer<LocalTime> withDateFormat(DateTimeFormatter dtf) {
+        return new LocalTimeDeserializer(dtf);
     }
 
     @Override